# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "GrantIn": {
        "additionalProperties": false,
        "properties": {
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "profileId": {
            "title": "Profileid",
            "type": "string"
          },
          "secretReference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secretreference"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "profileId",
          "permissions"
        ],
        "title": "GrantIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntegrationCreated": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "scopes",
          "token",
          "note"
        ],
        "title": "IntegrationCreated",
        "type": "object"
      },
      "IntegrationIn": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "default": "dataforge",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "IntegrationIn",
        "type": "object"
      },
...
      "ProfileIn": {
        "additionalProperties": false,
        "properties": {
          "defaultSchema": {
            "default": "",
            "title": "Defaultschema",
            "type": "string"
          },
          "environment": {
            "default": "development",
            "title": "Environment",
            "type": "string"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "mutatingRunbooksEnabled": {
            "default": false,
            "title": "Mutatingrunbooksenabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "port": {
            "default": 1521,
            "title": "Port",
            "type": "integer"
          },
          "protocol": {
            "default": "tcp",
            "title": "Protocol",
            "type": "string"
          },
          "secretReference": {
            "title": "Secretreference",
            "type": "string"
          },
          "serviceName": {
            "title": "Servicename",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "walletDir": {
            "default": "",
            "title": "Walletdir",
            "type": "string"
          },
          "worksheetsEnabled": {
            "default": false,
            "title": "Worksheetsenabled",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "host",
          "serviceName",
          "username",
          "secretReference"
        ],
        "title": "ProfileIn",
        "type": "object"
      },
...
      "SecretReferenceIn": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "default": "file",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "name",
          "locator"
        ],
        "title": "SecretReferenceIn",
        "type": "object"
      },
...
      "UserIn": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "default": "",
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "default": "",
            "title": "Email",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject"
        ],
        "title": "UserIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OracleDBHarness is the shared execution, policy and copilot layer for Oracle work.\n\nEvery operation carries an operation ID, an actor, a target, typed parameters,\ncapability requirements, a risk class and limits, and produces an execution record\nwith a structured result and verification evidence.\n\nNotes that apply to the whole API:\n\n* Free-form SQL runs only on targets where an administrator enabled worksheets. A\n  keyword filter is not a read-only boundary, so the Oracle account is the boundary.\n* Production targets are observation only in this release.\n* Applying a copilot proposal changes an editor buffer. It never runs, compiles or\n  commits anything.\n",
    "title": "OracleDBHarness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/grants": {
      "post": {
        "operationId": "create_grant_api_v1_admin_grants_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Grant Api V1 Admin Grants Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Grant",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/grants/{grant_id}": {
      "delete": {
        "operationId": "revoke_grant_api_v1_admin_grants__grant_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "grant_id",
            "required": true,
            "schema": {
              "title": "Grant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Revoke Grant Api V1 Admin Grants  Grant Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Grant",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/integrations": {
      "get": {
        "operationId": "list_integrations_api_v1_admin_integrations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Integrations Api V1 Admin Integrations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Integrations",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "description": "Issue a revocable integration credential.\n\nThe scope is deliberately narrow. ``copilot:assist`` lets an adapter ask for\nassistance; it grants no database execution capability and no harness target\naccess, and it does not make a DataForge administrator a harness administrator.",
        "operationId": "create_integration_api_v1_admin_integrations_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Integration",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/integrations/{integration_id}/revoke": {
      "post": {
        "operationId": "revoke_integration_api_v1_admin_integrations__integration_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "title": "Integration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Revoke Integration Api V1 Admin Integrations  Integration Id  Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Integration",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/secrets": {
      "get": {
        "operationId": "list_secrets_api_v1_admin_secrets_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Secrets Api V1 Admin Secrets Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Secrets",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "operationId": "create_secret_api_v1_admin_secrets_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretReferenceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Secret Api V1 Admin Secrets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Secret",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/targets": {
      "post": {
        "operationId": "create_target_api_v1_admin_targets_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Target Api V1 Admin Targets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Target",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Users Api V1 Admin Users Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Users",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_admin_users_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create User Api V1 Admin Users Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/targets": {
      "get": {
        "description": "Only targets the caller actually holds a grant on.",
        "operationId": "list_targets_api_v1_targets_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TargetView"
                  },
                  "title": "Response List Targets Api V1 Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Targets",
        "tags": [
          "targets"
        ]
      }
    },
...
  }
}